Refactor builder step to make it more modular
- Added FieldItem component for individual template fields with drag-and-drop support.
- Introduced IconColorPicker component for selecting icons and colors for templates.
- Created TemplateFieldsList component to manage and display the list of template fields.
- Developed useDragAndDrop hook to handle drag-and-drop logic for template fields.
- Implemented useScrollFade hook to manage scroll position and fade effects.
- Added constants for available icons and color options.
- Removed legacy template fields list implementation.
