+++ components/text-gen-builder/steps/filters.tsx
@@ -1,0 +1,2 @@
+// Re-export from modular structure
+export { default, validateRequiredFilters } from "./filters/index";
